reformat with Resharper

+++ Types/Bitbases.cs
@@ -4,13 +4,14 @@
 using ColorT = System.Int32;
 using SquareT = System.Int32;
 #endif
+
 internal static class Bitbases
 {
     // There are 24 possible pawn squares: the first 4 files and ranks from 2 to 7
-    private const int MAX_INDEX = 2*24*64*64; // stm * psq * wksq * bksq = 196608
+    private const int MAX_INDEX = 2 * 24 * 64 * 64; // stm * psq * wksq * bksq = 196608
 
     // Each uint32_t stores results of 32 positions, one per bit
-    internal static uint[] KPKBitbase = new uint[MAX_INDEX/32];
+    internal static uint[] KPKBitbase = new uint[MAX_INDEX / 32];
 
     // A KPK bitbase index is an integer in [0, IndexMax] range
     //
@@ -23,8 +24,8 @@ internal static class Bitbases
     // bit 15-17: white pawn RANK_7 - rank (from RANK_7 - RANK_7 to RANK_7 - RANK_2)
     internal static int index(ColorT us, SquareT bksq, SquareT wksq, SquareT psq)
     {
-        return (wksq | (bksq << 6) | (us << 12) | (Square.file_of(psq) << 13) |
-                 ((Rank.RANK_7 - Square.rank_of(psq)) << 15));
+        return (wksq | (bksq << 6) | (us << 12) | (Square.file_of(psq) << 13)
+                | ((Rank.RANK_7 - Square.rank_of(psq)) << 15));
     }
 
     internal static bool probe(SquareT wksq, SquareT wpsq, SquareT bksq, ColorT us)
@@ -32,7 +33,7 @@ internal static class Bitbases
         Debug.Assert(Square.file_of(wpsq) <= File.FILE_D);
 
         var idx = index(us, bksq, wksq, wpsq);
-        return (KPKBitbase[idx/32] & (1 << (int) (idx & 0x1F))) != 0;
+        return (KPKBitbase[idx / 32] & (1 << (idx & 0x1F))) != 0;
     }
 
     internal static void init()
@@ -61,7 +62,7 @@ internal static class Bitbases
         {
             if (db[idx] == Result.WIN)
             {
-                KPKBitbase[idx/32] |= (uint)(1 << (idx & 0x1F));
+                KPKBitbase[idx / 32] |= (uint)(1 << (idx & 0x1F));
             }
         }
     }
